web_platform/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gitpython>=3.1.40
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.24.0